refactor: tobe 패키지 구조화
