+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/handleradapter/HandlerAdapters.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.tobe.handleradapter;
 
 import java.util.ArrayList;
 import java.util.List;
