+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/handleradapter/ManualHandlerAdapter.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.tobe.handleradapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
