+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/handlermapping/HandlerMappings.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.tobe.handlermapping;
 
 import jakarta.servlet.http.HttpServletRequest;
